refactor python module structure - distribute core to different modules
