+++ .gitignore
@@ -144,6 +144,7 @@ docs/static/arrow/
 
 # Python
 *.pyc
+__pycache__/
 
 # Copied C++ source folders
 python/perspective/cmake
