+++ python/perspective/perspective/__init__.py
@@ -7,6 +7,10 @@
 #
 from .core import *  # noqa: F401, F403
 from .core._version import __version__  # noqa: F401
+from .manager import *  # noqa: F401, F403
+from .tornado_handler import *  # noqa: F401, F403
+from .viewer import *  # noqa: F401, F403
+from .widget import *  # noqa: F401, F403
 
 try:
     from .table import *  # noqa: F401, F403
