+++ python/perspective/perspective/manager/__init__.py
@@ -5,9 +5,7 @@
 # This file is part of the Perspective library, distributed under the terms of
 # the Apache License 2.0.  The full license can be found in the LICENSE file.
 #
-from ._version import __version__  # noqa: F401
-from .exception import PerspectiveError  # noqa: F401
-from .aggregate import Aggregate  # noqa: F401
-from .filters import ALL_FILTERS  # noqa: F401
-from .plugin import Plugin  # noqa: F401
-from .sort import Sort  # noqa: F401
+from .manager import PerspectiveManager  # noqa: F401
+from .session import PerspectiveSession  # noqa: F401
+
+__all__ = ["PerspectiveManager", "PerspectiveSession"]
