+++ python/perspective/perspective/manager/session.py
